Queuify observation plans
