+++ services/observation_plan_queue/supervisor.conf
@@ -1,0 +1,5 @@
+[program:observation_plan_queue]
+command=/usr/bin/env python services/observation_plan_queue/observation_plan_queue.py %(ENV_FLAGS)s
+environment=PYTHONPATH=".",PYTHONUNBUFFERED="1"
+stdout_logfile=log/observation_plan.log
+redirect_stderr=true
