+++ skyportal/handlers/api/gcn.py
@@ -20,6 +20,7 @@ from tornado.ioloop import IOLoop
 import arrow
 import astropy
 import humanize
+import requests
 import sqlalchemy as sa
 from sqlalchemy.orm import joinedload
 from sqlalchemy.orm import sessionmaker, scoped_session
@@ -46,16 +47,11 @@ from baselayer.app.env import load_env
 from baselayer.app.flow import Flow
 
 from .source import post_source
-from .observation_plan import (
-    post_observation_plan,
-    post_survey_efficiency_analysis,
-)
 from ..base import BaseHandler
 from ...models import (
     DBSession,
     Allocation,
     CatalogQuery,
-    DefaultObservationPlanRequest,
     EventObservationPlan,
     GcnEvent,
     GcnNotice,
@@ -1424,167 +1420,28 @@ def add_tiles_and_properties_and_contour(localization_id, user_id, parent_sessio
             Session.remove()
 
 
-def add_observation_plans(localization_id, user_id, parent_session=None):
+def add_observation_plans(localization_id, user_id):
 
-    if parent_session is None:
-        if Session.registry.has():
-            session = Session()
-        else:
-            session = Session(bind=DBSession.session_factory.kw["bind"])
-    else:
-        session = parent_session
+    request_body = {
+        'localization_id': localization_id,
+        'user_id': user_id,
+    }
 
-    try:
-        user = session.scalar(sa.select(User).where(User.id == user_id))
-        localization = session.query(Localization).get(localization_id)
-        localization_tags = [
-            tags.text
-            for tags in session.query(LocalizationTag)
-            .filter(LocalizationTag.localization_id == localization_id)
-            .all()
-        ]
-        dateobs = localization.dateobs
-        config_gcn_observation_plans_all = [
-            observation_plan for observation_plan in cfg["gcn.observation_plans"]
-        ]
-        config_gcn_observation_plans = []
-        for config_gcn_observation_plan in config_gcn_observation_plans_all:
-            allocation = session.scalars(
-                Allocation.select(user).where(
-                    Allocation.proposal_id
-                    == config_gcn_observation_plan["allocation-proposal_id"]
-                )
-            ).first()
-            if allocation is not None:
-                allocation_id = allocation.id
-                config_gcn_observation_plan["allocation_id"] = allocation_id
-                config_gcn_observation_plan["survey_efficiencies"] = []
-                config_gcn_observation_plans.append(config_gcn_observation_plan)
-            else:
-                allocation_id = None
+    observation_plans_microservice_url = (
+        f'http://127.0.0.1:{cfg["ports.observation_plans_queue"]}'
+    )
 
-        default_observation_plans = (
-            (
-                session.scalars(
-                    DefaultObservationPlanRequest.select(
-                        user,
-                        options=[
-                            joinedload(
-                                DefaultObservationPlanRequest.default_survey_efficiencies
-                            )
-                        ],
-                    )
-                )
-            )
-            .unique()
-            .all()
+    resp = requests.post(
+        observation_plans_microservice_url, json=request_body, timeout=2
+    )
+    if resp.status_code == 200:
+        log(
+            f'Observation plans requested for Localization with ID {request_body["localization_id"]}'
         )
-        gcn_observation_plans = []
-        for plan in default_observation_plans:
-            allocation = session.scalars(
-                Allocation.select(user).where(Allocation.id == plan.allocation_id)
-            ).first()
-
-            gcn_observation_plan = {
-                'allocation_id': allocation_id,
-                'filters': plan.filters,
-                'payload': plan.payload,
-                'survey_efficiencies': [
-                    survey_efficiency.to_dict()
-                    for survey_efficiency in plan.default_survey_efficiencies
-                ],
-            }
-            gcn_observation_plans.append(gcn_observation_plan)
-        gcn_observation_plans = gcn_observation_plans + config_gcn_observation_plans
-
-        event = session.scalars(
-            GcnEvent.select(user).where(GcnEvent.dateobs == dateobs)
-        ).first()
-        start_date = str(datetime.datetime.utcnow()).replace("T", "")
-
-        for ii, gcn_observation_plan in enumerate(gcn_observation_plans):
-            allocation_id = gcn_observation_plan['allocation_id']
-            allocation = session.scalars(
-                Allocation.select(user).where(Allocation.id == allocation_id)
-            ).first()
-
-            if allocation is not None:
-
-                end_date = allocation.instrument.telescope.next_sunrise()
-                if end_date is None:
-                    end_date = str(
-                        datetime.datetime.utcnow() + datetime.timedelta(days=1)
-                    ).replace("T", "")
-                else:
-                    end_date = Time(end_date, format='jd').iso
-
-                payload = {
-                    **gcn_observation_plan['payload'],
-                    'start_date': start_date,
-                    'end_date': end_date,
-                    'queue_name': f'{allocation.instrument.name}-{start_date}-{ii}',
-                }
-                plan = {
-                    'payload': payload,
-                    'allocation_id': allocation.id,
-                    'gcnevent_id': event.id,
-                    'localization_id': localization_id,
-                }
-
-                if 'filters' in gcn_observation_plan:
-                    filters = gcn_observation_plan['filters']
-                    if filters is not None:
-                        if 'gcn_notices' in filters and len(filters['gcn_notices']) > 0:
-                            if not any(
-                                [
-                                    gcn.NoticeType(notice.notice_type).name
-                                    in filters['gcn_notices']
-                                    for notice in event.gcn_notices
-                                ]
-                            ):
-                                continue
-
-                        if 'gcn_tags' in filters and len(filters['gcn_tags']) > 0:
-                            intersection = list(
-                                set(event.tags) & set(filters["gcn_tags"])
-                            )
-                            if len(intersection) == 0:
-                                continue
-
-                        if (
-                            "localization_tags" in filters
-                            and len(filters["localization_tags"]) > 0
-                        ):
-                            intersection = list(
-                                set(localization_tags)
-                                & set(filters["localization_tags"])
-                            )
-                            if len(intersection) == 0:
-                                continue
-
-                plan_id = post_observation_plan(
-                    plan, user_id, session, asynchronous=False
-                )
-                for survey_efficiency in gcn_observation_plan['survey_efficiencies']:
-                    try:
-                        post_survey_efficiency_analysis(
-                            survey_efficiency,
-                            plan_id,
-                            user_id,
-                            session,
-                            asynchronous=False,
-                        )
-                    except Exception as e:
-                        log(f"Survey efficiency analysis failed: {str(e)}")
-                        continue
-
-        return log(f"Generated observation plans for localization {localization_id}")
-    except Exception as e:
-        log(f"Unable to observation plans for localization {localization_id}: {e}")
-    finally:
-        if parent_session is None:
-            session.close()
-            Session.remove()
+    else:
+        log(
+            f'Observation plan request failed for Localization with ID {request_body["localization_id"]}: {resp.content}'
+        )
 
 
 def add_tiles_properties_contour_and_obsplan(
@@ -1601,7 +1458,7 @@ def add_tiles_properties_contour_and_obsplan(
 
     try:
         add_tiles_and_properties_and_contour(localization_id, user_id, session)
-        add_observation_plans(localization_id, user_id, session)
+        add_observation_plans(localization_id, user_id)
     except Exception as e:
         log(
             f"Unable to generate tiles / properties / observation plans / contour for localization {localization_id}: {e}"
