+++ skyportal/services/test_server/supervisor.conf
@@ -39,3 +39,9 @@ command=/usr/bin/env python services/notification_queue/notification_queue.py %(
 environment=PYTHONPATH=".",PYTHONUNBUFFERED="1"
 stdout_logfile=log/notification.log
 redirect_stderr=true
+
+[program:observation_plans_queue]
+command=/usr/bin/env python services/observation_plans_queue/observation_plans_queue.py %(ENV_FLAGS)s
+environment=PYTHONPATH=".",PYTHONUNBUFFERED="1"
+stdout_logfile=log/observation_plans.log
+redirect_stderr=true
